FlaskProject_server/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
openai==1.68.2
langchain==0.3.23
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.26
html2text==2024.2.26
requests==2.32.3
httpx==0.28.1
pydantic==2.10.6
tiktoken==0.9.0
pymongo==4.8.0
langdetect==1.0.9
redis>=5.0.0
python-dotenv>=1.0.1
bcrypt>=4.0.1
beautifulsoup4==4.13.3
nest_asyncio>=1.5.1
numpy==1.24.3
faiss-cpu==1.7.4